Remove unnecessary DividerItemDecoration class Refactor code in agenda package
