+++ app/src/main/java/com/arthurnagy/droidconberlin/feature/agenda/MyAgendaFragment.kt
@@ -2,14 +2,14 @@ package com.arthurnagy.droidconberlin.feature.agenda
 
 import android.databinding.DataBindingUtil
 import android.os.Bundle
+import android.support.v7.widget.DividerItemDecoration
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.arthurnagy.droidconberlin.MyAgendaBinding
 import com.arthurnagy.droidconberlin.R
 import com.arthurnagy.droidconberlin.architecture.DroidconFragment
-import com.arthurnagy.droidconberlin.dimension
-import com.arthurnagy.droidconberlin.feature.shared.DividerItemDecoration
+import com.arthurnagy.droidconberlin.feature.shared.StickyHeaderItemDecoration
 import com.arthurnagy.droidconberlin.setupToolbar
 
 class MyAgendaFragment : DroidconFragment() {
@@ -26,9 +26,11 @@ class MyAgendaFragment : DroidconFragment() {
         super.onViewCreated(view, savedInstanceState)
         setupToolbar(binding.toolbar)
         binding.viewModel = viewModel
-        binding.recyclerView.addItemDecoration(StickyHeaderItemDecoration(context,
-                isHeader = { position -> position % MyAgendaViewModel.ITEM_COUNT_PER_CATEGORY == 0 },
-                getHeaderTitle = { position -> "Category ${position / MyAgendaViewModel.ITEM_COUNT_PER_CATEGORY + 1}" }))
-        binding.recyclerView.addItemDecoration(DividerItemDecoration(context))
+        binding.recyclerView.addItemDecoration(object : StickyHeaderItemDecoration(context) {
+            override fun isHeader(position: Int) = position % MyAgendaViewModel.ITEM_COUNT_PER_CATEGORY == 0
+
+            override fun getHeaderTitle(position: Int) = "Category ${position / MyAgendaViewModel.ITEM_COUNT_PER_CATEGORY + 1}"
+        })
+        binding.recyclerView.addItemDecoration(DividerItemDecoration(context, DividerItemDecoration.VERTICAL))
     }
 }
